+++ src/nix/mod.rs
@@ -1,2 +1,3 @@
 pub mod devour_flake;
 pub mod lock;
+pub mod system_list;
